+++ Cosette/Engine/Ai/Score/PieceSquareTables/BishopTables.cs
@@ -32,15 +32,15 @@ namespace Cosette.Engine.Ai.Score.PieceSquareTables
             // White
             new []
             {
-                TableOperations.FlipVertically(Pattern[(int)GamePhase.Opening]),
-                TableOperations.FlipVertically(Pattern[(int)GamePhase.Ending])
+                TableOperations.FlipVertically(Pattern[GamePhase.Opening]),
+                TableOperations.FlipVertically(Pattern[GamePhase.Ending])
             },
 
             // Black
             new []
             {
-                TableOperations.FlipHorizontally(Pattern[(int)GamePhase.Opening]),
-                TableOperations.FlipHorizontally(Pattern[(int)GamePhase.Ending])
+                TableOperations.FlipHorizontally(Pattern[GamePhase.Opening]),
+                TableOperations.FlipHorizontally(Pattern[GamePhase.Ending])
             }
         };
     }
