+++ Cosette/Engine/Perft/VerificationPerft.cs
@@ -71,10 +71,10 @@ namespace Cosette.Engine.Perft
                 return false;
             }
 
-            if (board.Position[Color.White][(int)GamePhase.Opening] != board.CalculatePosition(Color.White, GamePhase.Opening) ||
-                board.Position[Color.White][(int)GamePhase.Ending] != board.CalculatePosition(Color.White, GamePhase.Ending) ||
-                board.Position[Color.Black][(int)GamePhase.Opening] != board.CalculatePosition(Color.Black, GamePhase.Opening) ||
-                board.Position[Color.Black][(int)GamePhase.Ending] != board.CalculatePosition(Color.Black, GamePhase.Ending))
+            if (board.Position[Color.White][GamePhase.Opening] != board.CalculatePosition(Color.White, GamePhase.Opening) ||
+                board.Position[Color.White][GamePhase.Ending] != board.CalculatePosition(Color.White, GamePhase.Ending) ||
+                board.Position[Color.Black][GamePhase.Opening] != board.CalculatePosition(Color.Black, GamePhase.Opening) ||
+                board.Position[Color.Black][GamePhase.Ending] != board.CalculatePosition(Color.Black, GamePhase.Ending))
             {
                 return false;
             }
